# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Seoul Subway API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Station {
  id: string;
  name: string;
  line: string;
  lineColor: string;
  lineNumber: string;
  transferLines?: string[];
}

export interface SubwayLine {
  id: string;
  name: string;
  shortName: string;
  color: string;
}

export interface ArrivalInfo {
  trainNo: string;
  destination: string;
  remainingTime: string;
  status: string;
  line: string;
  lineColor?: string;
  currentStation: string;
  direction?: string;
  isLastTrain?: boolean;
}

export interface ArrivalResponse {
  stationName: string;
  arrivals: ArrivalInfo[];
  updatedAt?: string;
}

export interface RouteSegment {
  stations: string[];
  line: string;
  lineColor: string;
  lineNumber: string;
  boardingInfo?: string;
}

export interface RouteResult {
  from: string;
  to: string;
  segments: RouteSegment[];
  totalStations: number;
  transferCount: number;
  estimatedTime?: number;
}

export interface StationFacility {
  type: string;
  location: string;
  description: string;
}

export interface StationExit {
  number: string;
  description: string;
}

export interface StationMap {
  stationName: string;
  floors: string[];
  exits: StationExit[];
  facilities: StationFacility[];
  /** @nullable */
  transferInfo?: string | null;
}

export interface IndoorFloor {
  level: string;
  name: string;
}

export interface IndoorExit {
  number: string;
  description: string;
}

export interface IndoorFacility {
  type: string;
  location: string;
  floors: string;
}

export interface IndoorDirection {
  toward: string;
  side: string;
}

export interface IndoorTransfer {
  line: string;
  floor: string;
  description: string;
}

export type IndoorNavResponseAllLinesItem = {
  line?: string;
  voiceGuide?: string;
};

export interface IndoorNavResponse {
  stationCode: string;
  stationName: string;
  line: string;
  floors: IndoorFloor[];
  exits: IndoorExit[];
  facilities: IndoorFacility[];
  directions: IndoorDirection[];
  transfers: IndoorTransfer[];
  voiceGuide: string;
  /** All results for transfer stations (multiple lines at same station) */
  allLines?: IndoorNavResponseAllLinesItem[];
}

export interface ErrorResponse {
  error: string;
}

export type SearchStationsParams = {
  /**
   * Search query (station name)
   */
  q?: string;
  /**
   * Filter by line number (e.g. "1", "2", "경의중앙")
   */
  line?: string;
};

export type FindRouteParams = {
  /**
   * Departure station name
   */
  from: string;
  /**
   * Destination station name
   */
  to: string;
};

export type GetIndoorNavParams = {
  /**
   * Line number to disambiguate transfer stations (e.g. "1", "2")
   */
  line?: string;
};